+++ languages/de/explain.php
@@ -17,7 +17,7 @@ $GLOBALS['TL_LANG']['XPL']['anystores_apiKey'] = array
     array
     (
         'colspan',
-        'Bitte beachten Sie, dass Sie den Key für folgende APIs freischalten müssen.<br>'.
+        'Bitte beachten Sie, dass Sie den <b>Server-Key</b> für folgende APIs freischalten müssen.<br>'.
         'Mehr Informationen finden Sie hier: <a href="https://console.developers.google.com" target="_blank">https://console.developers.google.com</a>'
     ),
     array
@@ -27,13 +27,26 @@ $GLOBALS['TL_LANG']['XPL']['anystores_apiKey'] = array
     ),
     array
     (
-        'Google Maps Geolocation API',
-        'Zum Berechnen der Geokoordinaten aus einer Adresse.'
+        'Google Static Maps API',
+        'Zum Rendern der statischen Karte im Backend.'
+    )
+
+);
+
+
+$GLOBALS['TL_LANG']['XPL']['anystores_apiBrowserKey'] = array
+(
+
+    array
+    (
+        'colspan',
+        'Bitte beachten Sie, dass Sie den <b>Browser-Key</b> für folgende APIs freischalten müssen.<br>'.
+        'Mehr Informationen finden Sie hier: <a href="https://console.developers.google.com" target="_blank">https://console.developers.google.com</a>'
     ),
     array
     (
-        'Google Static Maps API',
-        'Zum Rendern der statischen Karte im Backend.'
+        'Google Maps JavaScript API',
+        'Zum Anzeigen der dynamischen Karte.'
     )
 
 );