+++ languages/de/tl_anystores_settings.php
@@ -20,6 +20,7 @@ $GLOBALS['TL_LANG']['tl_anystores_settings'] = array
     // Fields
     'anystores_defaultMarker' => array('Standard-Marker für alle Standorte', 'Kann im Modul, in der Kategorie und am Standort überschrieben werden.'),
     'anystores_geoApi'        => array('API-Anbieter', 'Wählen Sie einen API-Anbieter aus.'),
-    'anystores_apiKey'        => array('API-Key', 'Geben Sie Ihren API-Key ein.')
+    'anystores_apiKey'        => array('Server-API-Key', 'Geben Sie Ihren Server-API-Key ein.'),
+    'anystores_apiBrowserKey' => array('Browser-API-Key', 'Geben Sie Ihren Browser-API-Key ein.'),
 
 );