+++ languages/en/explain.php
@@ -17,7 +17,7 @@ $GLOBALS['TL_LANG']['XPL']['anystores_apiKey'] = array
     array
     (
         'colspan',
-        'Please make sure you’re enable the following APIs.<br>'.
+        'Please make sure you’re enable for the <b>server key</b> the following APIs.<br>'.
         'More informations: <a href="https://console.developers.google.com" target="_blank">https://console.developers.google.com</a>'
     ),
     array
@@ -29,6 +29,23 @@ $GLOBALS['TL_LANG']['XPL']['anystores_apiKey'] = array
     (
         'Google Static Maps API',
         'To render the static maps in backend.'
+    )
+
+);
+
+$GLOBALS['TL_LANG']['XPL']['anystores_apiBrowserKey'] = array
+(
+
+    array
+    (
+        'colspan',
+        'Please make sure you’re enable for the <b>browser key</b> the following APIs.<br>'.
+        'More informations: <a href="https://console.developers.google.com" target="_blank">https://console.developers.google.com</a>'
+    ),
+    array
+    (
+        'Google Maps JavaScript API',
+        'Get the geo coordinates from locations.'
     )
 
 );