+++ models/AnyStoresModel.php
@@ -3,7 +3,7 @@
 /**
  * anyStores for Contao Open Source CMS
  *
- * @copyright   (c) 2014, 2015 Tastaturberuf <mail@tastaturberuf.de>
+ * @copyright   (c) 2014 - 2016 Tastaturberuf <mail@tastaturberuf.de>
  * @author      Daniel Jahnsmüller <mail@jahnsmueller.net>
  * @license     http://opensource.org/licenses/lgpl-3.0.html
  * @package     anyStores
@@ -31,13 +31,22 @@ class AnyStoresModel extends \Model
 
     /**
      * Find all published
+     *
+     * @param array $arrOptions
+     *
+     * @return \Model\Collection|null
      */
-    public static function findAllPublished()
+    public static function findAllPublished(array $arrOptions = array())
     {
-        $arrOptions = array
+        $arrOptions = array_merge
         (
-            'column' => 'published',
-            'value'  => 1
+            array
+            (
+                'column' => 'published',
+                'value'  => 1
+            ),
+
+            $arrOptions
         );
 
         return static::findAll($arrOptions);